Refactored AttributeView()
