+++ src/compas/datastructures/mesh/core/__init__.py
@@ -4,6 +4,7 @@ from __future__ import print_function
 
 from compas import IPY
 
+from .attributes import *  # noqa: F401 F403
 from .halfedge import HalfEdge  # noqa: F401
 from .mesh import BaseMesh  # noqa: F401
 from .operations import *  # noqa: F401 F403
